chore(board): move utils file
