+++ next-tavla/src/Board/scenarios/Board/utils.ts
